+++ src/carrossel.js
@@ -1,0 +1,58 @@
+const fila = document.querySelector('.container-carrossel');
+const pokemons = document.querySelectorAll('.pokemon');
+const flechaEsquerda = document.getElementById('flecha-esquerda');
+const flechaDireita = document.getElementById('flecha-direita');
+
+//event listener para a flecha direita
+flechaDireita.addEventListener('click', () => {
+    fila.scrollLeft += fila.offsetWidth;
+    const indicadorActivo = document.querySelector('.indicadores .activo');
+    if(indicadorActivo.nextSibling){
+        indicadorActivo.nextSibling.classList.add('activo');
+        indicadorActivo.classList.remove('activo');
+    }
+});
+
+//event listener para a flecha esquerda
+flechaEsquerda.addEventListener('click', () => {
+    fila.scrollLeft -= fila.offsetWidth;
+    const indicadorActivo = document.querySelector('.indicadores .activo');
+    if(indicadorActivo.previousSibling){
+        indicadorActivo.previousSibling.classList.add('activo');
+        indicadorActivo.classList.remove('activo');
+    }
+});
+
+//paginação
+const numeroPaginas = Math.ceil(pokemons.length / 5);
+for(let i = 0; i < numeroPaginas; i++){
+    const indicador = document.createElement('button');
+
+    if(i === 0){
+        indicador.classList.add('activo');
+    }
+
+    document.querySelector('.indicadores').appendChild(indicador);
+    indicador.addEventListener('click', (e) => {
+        fila.scrollLeft = i * fila.offsetWidth;
+
+        document.querySelector('.indicadores .activo').classList.remove('activo');
+        e.target.classList.add('activo');
+    });
+
+}
+
+//hover
+pokemons.forEach((pokemon) => {
+    pokemon.addEventListener('mouseenter', (e) => {
+        const elemento = e.currentTarget;
+        setTimeout(() => {
+        pokemons.forEach(pokemon => pokemon.classList.remove('hover'));
+        elemento.classList.add('hover');
+        }, 300);
+    });
+});
+
+fila.addEventListener('mouseleave', () => {
+    pokemons.forEach(pokemon => pokemon.classList.remove('hover'));
+});
