+++ src/data.js
@@ -1,9 +1,10 @@
-// estas funciones son de ejemplo
-
 export const example = () => {
   return 'example';
 };
 
 export const anotherExample = () => {
+
   return 'OMG';
 };
+
+//logica aqui
