+++ src/index.html
@@ -3,7 +3,7 @@
 
 <head>
 	<meta charset="utf-8" />
-	<meta name="description" content="Cards de Pokemons">
+	<meta name="description" content="Pokemons">
 	<meta name="language" content="pt-br" />
 	<link rel="shortcut icon" href="img/favicon.ico" />
 	<link rel="preconnect" href="https://fonts.gstatic.com">
@@ -19,9 +19,9 @@
       <a ref=""><img class="logo-tipo" src="img/logo.png" alt="logo" /></a>
       <nav>
         <div class="menu">
-        <a href="#" class="ativo">Início</a>
-        <a href="#">Estatísticas</a>
-        <a href="#">Sobre</a>
+          <a href="#" class="ativo">Início</a>
+          <a href="#">Estatísticas</a>
+          <a href="#">Sobre</a>
         </div>
       </nav>
     </div>
@@ -31,17 +31,18 @@
       <div class="container">
         <h2 class="titulo">PokeInfos </h2>
           <p class="descricao">
-            *****ESCREVER ALGO******
+            Bem vindo ao PokeInfo, sua central de informações Pokémon! 
           </p>
       </div>
     </div>
     <div class="pokemons-recomendados container">
-      <div class="container-title-control">
-        <h3 class="titulo-carrossel">Pokemóns de Kanto </h3>
-        <div class="indicadores">
+
+      <div>
+        <div class="container-title-control">
+          <h3 class="titulo-carrossel">Pokemóns de Kanto </h3>
+          <div class="indicadores"></div>
         </div>
-      </div>
-      <div class="container-principal">
+        <div class="container-principal">
         <button role="button" id="flecha-esquerda" class="flecha-esquerda"><i class="fas fa-chevron-left"></i></button>
           
         <div class="container-carrossel">
@@ -91,10 +92,126 @@
         </div>
           
         <button role="button" id="flecha-direita" class="flecha-direita"><i class="fas fa-chevron-right"></i></button>
+        </div>
+      <!-- </div>
+      <div>
+        <div class="container-title-control">
+          <h3 class="titulo-carrossel">Pokemóns de Johto</h3>
+          <div class="indicadores"></div>
+        </div>
+        <div class="container-principal">
+        <button role="button" id="flecha-esquerda" class="flecha-esquerda"><i class="fas fa-chevron-left"></i></button>
+          
+        <div class="container-carrossel">
+          <div class="carrossel">
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/001.png" alt="bulbasaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/002.png" alt="ivysaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/003.png" alt="venusaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/004.png" alt="charmander"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/005.png" alt="charmeleon"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/006.png" alt="charizard"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/007.png" alt="squirtle"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/001.png" alt="bulbasaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/002.png" alt="ivysaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/003.png" alt="venusaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/004.png" alt="charmander"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/005.png" alt="charmeleon"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/006.png" alt="charizard"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/007.png" alt="squirtle"></a>
+            </div>
+        </div>
+          
+        <button role="button" id="flecha-direita" class="flecha-direita"><i class="fas fa-chevron-right"></i></button>
+        </div>
+      </div>
+      <div>
+        <div class="container-title-control">
+          <h3 class="titulo-carrossel">Pokemóns de Kanto </h3>
+          <div class="indicadores"></div>
+        </div>
+        <div class="container-principal">
+        <button role="button" id="flecha-esquerda" class="flecha-esquerda"><i class="fas fa-chevron-left"></i></button>
+          
+        <div class="container-carrossel">
+          <div class="carrossel">
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/001.png" alt="bulbasaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/002.png" alt="ivysaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/003.png" alt="venusaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/004.png" alt="charmander"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/005.png" alt="charmeleon"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/006.png" alt="charizard"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/007.png" alt="squirtle"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/001.png" alt="bulbasaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/002.png" alt="ivysaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/003.png" alt="venusaur"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/004.png" alt="charmander"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/005.png" alt="charmeleon"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/006.png" alt="charizard"></a>
+            </div>
+            <div class="pokemon">
+            <a href="#"><img src="https://www.serebii.net/pokemongo/pokemon/007.png" alt="squirtle"></a>
+            </div>
+        </div>
+          
+        <button role="button" id="flecha-direita" class="flecha-direita"><i class="fas fa-chevron-right"></i></button>
+        </div> -->
       </div>
     </div>
   </main>
   <script src="https://kit.fontawesome.com/2c36e9b7b1.js" crossorigin="anonymous"></script>
+  <script src="carrossel.js" type="module"></script>
   <script src="main.js" type="module"></script>
   </body>
 </html>
